pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.7"
fastapi = "^0.111.0"
uvicorn = {extras = ["standart"], version = "^0.30.0"}
pydantic = {extras = ["email"], version = "^2.7.2"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
jinja2 = "^3.1.4"
psycopg2 = "^2.9.9"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"